requirements.txt
langchain_openai 
langchain_community 
langchain 
pymysql 
chromadb
pandas
django
django_rest_framework
python-dotenv
